requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pymongo>=4.6.0
python-dotenv>=1.0.0
schedule>=1.2.0
lxml>=5.1.0
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.1.0
